KAFKA-14966 - Extract reusable functionality from OffsetFetcher
Initial refactoring for extracting reusable logic from OffsetFetcher into OffsetFetcherUtils. No changes in logic, just moving classes, functions or pieces of logic into OffsetFetcherUtils.
All the functionality moved from OffsetFetcher to the OffsetFetcherUtils is already covered by tests in OffsetFetcherTest and FetcherTest. There were no individual tests for the extracted functions, so no tests were migrated.
