+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/OffsetFetcher.java
@@ -39,18 +39,16 @@ import org.apache.kafka.clients.consumer.OffsetAndTimestamp;
 import org.apache.kafka.clients.consumer.OffsetResetStrategy;
 import org.apache.kafka.clients.consumer.internals.OffsetsForLeaderEpochClient.OffsetForEpochResult;
 import org.apache.kafka.clients.consumer.internals.SubscriptionState.FetchPosition;
+import org.apache.kafka.clients.consumer.internals.OffsetFetcherUtils.ListOffsetResult;
+import org.apache.kafka.clients.consumer.internals.OffsetFetcherUtils.ListOffsetData;
 import org.apache.kafka.common.IsolationLevel;
 import org.apache.kafka.common.Node;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.errors.RetriableException;
 import org.apache.kafka.common.errors.TimeoutException;
-import org.apache.kafka.common.errors.TopicAuthorizationException;
 import org.apache.kafka.common.message.ApiVersionsResponseData.ApiVersion;
 import org.apache.kafka.common.message.ListOffsetsRequestData.ListOffsetsPartition;
-import org.apache.kafka.common.message.ListOffsetsResponseData.ListOffsetsPartitionResponse;
-import org.apache.kafka.common.message.ListOffsetsResponseData.ListOffsetsTopicResponse;
 import org.apache.kafka.common.protocol.ApiKeys;
-import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.requests.ListOffsetsRequest;
 import org.apache.kafka.common.requests.ListOffsetsResponse;
 import org.apache.kafka.common.requests.OffsetsForLeaderEpochRequest;
@@ -75,10 +73,9 @@ public class OffsetFetcher {
     private final long requestTimeoutMs;
     private final IsolationLevel isolationLevel;
     private final AtomicReference<RuntimeException> cachedListOffsetsException = new AtomicReference<>();
-    private final AtomicReference<RuntimeException> cachedOffsetForLeaderException = new AtomicReference<>();
     private final OffsetsForLeaderEpochClient offsetsForLeaderEpochClient;
     private final ApiVersions apiVersions;
-    private final AtomicInteger metadataUpdateVersion = new AtomicInteger(-1);
+    private final OffsetFetcherUtils offsetFetcherUtils;
 
     public OffsetFetcher(LogContext logContext,
                          ConsumerNetworkClient client,
@@ -99,31 +96,8 @@ public class OffsetFetcher {
         this.isolationLevel = isolationLevel;
         this.apiVersions = apiVersions;
         this.offsetsForLeaderEpochClient = new OffsetsForLeaderEpochClient(client, logContext);
-    }
-
-    /**
-     * Represents data about an offset returned by a broker.
-     */
-    static class ListOffsetData {
-        final long offset;
-        final Long timestamp; //  null if the broker does not support returning timestamps
-        final Optional<Integer> leaderEpoch; // empty if the leader epoch is not known
-
-        ListOffsetData(long offset, Long timestamp, Optional<Integer> leaderEpoch) {
-            this.offset = offset;
-            this.timestamp = timestamp;
-            this.leaderEpoch = leaderEpoch;
-        }
-    }
-
-    private Long offsetResetStrategyTimestamp(final TopicPartition partition) {
-        OffsetResetStrategy strategy = subscriptions.resetStrategy(partition);
-        if (strategy == OffsetResetStrategy.EARLIEST)
-            return ListOffsetsRequest.EARLIEST_TIMESTAMP;
-        else if (strategy == OffsetResetStrategy.LATEST)
-            return ListOffsetsRequest.LATEST_TIMESTAMP;
-        else
-            return null;
+        this.offsetFetcherUtils = new OffsetFetcherUtils(logContext, metadata, subscriptions,
+            time, apiVersions);
     }
 
     private OffsetResetStrategy timestampToOffsetResetStrategy(long timestamp) {
@@ -147,16 +121,11 @@ public class OffsetFetcher {
         if (exception != null)
             throw exception;
 
-        Set<TopicPartition> partitions = subscriptions.partitionsNeedingReset(time.milliseconds());
-        if (partitions.isEmpty())
-            return;
+        Map<TopicPartition, Long> offsetResetTimestamps =
+            offsetFetcherUtils.getOffsetResetTimestamp();
 
-        final Map<TopicPartition, Long> offsetResetTimestamps = new HashMap<>();
-        for (final TopicPartition partition : partitions) {
-            Long timestamp = offsetResetStrategyTimestamp(partition);
-            if (timestamp != null)
-                offsetResetTimestamps.put(partition, timestamp);
-        }
+        if (offsetResetTimestamps.isEmpty())
+            return;
 
         resetPositionsAsync(offsetResetTimestamps);
     }
@@ -165,27 +134,15 @@ public class OffsetFetcher {
      * Validate offsets for all assigned partitions for which a leader change has been detected.
      */
     public void validatePositionsIfNeeded() {
-        RuntimeException exception = cachedOffsetForLeaderException.getAndSet(null);
-        if (exception != null)
-            throw exception;
-
-        // Validate each partition against the current leader and epoch
-        // If we see a new metadata version, check all partitions
-        validatePositionsOnMetadataChange();
-
-        // Collect positions needing validation, with backoff
-        Map<TopicPartition, FetchPosition> partitionsToValidate = subscriptions
-                .partitionsNeedingValidation(time.milliseconds())
-                .stream()
-                .filter(tp -> subscriptions.position(tp) != null)
-                .collect(Collectors.toMap(Function.identity(), subscriptions::position));
+        Map<TopicPartition, SubscriptionState.FetchPosition> partitionsToValidate =
+            offsetFetcherUtils.getPartitionsToValidate();
 
         validatePositionsAsync(partitionsToValidate);
     }
 
     public Map<TopicPartition, OffsetAndTimestamp> offsetsForTimes(Map<TopicPartition, Long> timestampsToSearch,
                                                                    Timer timer) {
-        metadata.addTransientTopics(topicsForPartitions(timestampsToSearch.keySet()));
+        metadata.addTransientTopics(offsetFetcherUtils.topicsForPartitions(timestampsToSearch.keySet()));
 
         try {
             Map<TopicPartition, ListOffsetData> fetchedOffsets = fetchOffsetsByTimes(timestampsToSearch,
@@ -287,7 +244,7 @@ public class OffsetFetcher {
     private Map<TopicPartition, Long> beginningOrEndOffset(Collection<TopicPartition> partitions,
                                                            long timestamp,
                                                            Timer timer) {
-        metadata.addTransientTopics(topicsForPartitions(partitions));
+        metadata.addTransientTopics(offsetFetcherUtils.topicsForPartitions(partitions));
         try {
             Map<TopicPartition, Long> timestampsToSearch = partitions.stream()
                     .distinct()
@@ -420,7 +377,7 @@ public class OffsetFetcher {
                     });
 
                     if (!truncations.isEmpty()) {
-                        maybeSetOffsetForLeaderException(buildLogTruncationException(truncations));
+                        offsetFetcherUtils.maybeSetOffsetForLeaderException(buildLogTruncationException(truncations));
                     }
                 }
 
@@ -430,7 +387,7 @@ public class OffsetFetcher {
                     metadata.requestUpdate();
 
                     if (!(e instanceof RetriableException)) {
-                        maybeSetOffsetForLeaderException(e);
+                        offsetFetcherUtils.maybeSetOffsetForLeaderException(e);
                     }
                 }
             });
@@ -447,12 +404,6 @@ public class OffsetFetcher {
         }
         return new LogTruncationException("Detected truncated partitions: " + truncations,
                 truncatedFetchOffsets, divergentOffsets);
-    }
-
-    private void maybeSetOffsetForLeaderException(RuntimeException e) {
-        if (!cachedOffsetForLeaderException.compareAndSet(null, e)) {
-            log.error("Discarding error in OffsetsForLeaderEpoch because another error is pending", e);
-        }
     }
 
     /**
@@ -544,7 +495,7 @@ public class OffsetFetcher {
                 }
             }
         }
-        return regroupPartitionMapByNode(partitionDataMap);
+        return offsetFetcherUtils.regroupPartitionMapByNode(partitionDataMap);
     }
 
     /**
@@ -586,112 +537,22 @@ public class OffsetFetcher {
      */
     private void handleListOffsetResponse(ListOffsetsResponse listOffsetsResponse,
                                           RequestFuture<ListOffsetResult> future) {
-        Map<TopicPartition, ListOffsetData> fetchedOffsets = new HashMap<>();
-        Set<TopicPartition> partitionsToRetry = new HashSet<>();
-        Set<String> unauthorizedTopics = new HashSet<>();
-
-        for (ListOffsetsTopicResponse topic : listOffsetsResponse.topics()) {
-            for (ListOffsetsPartitionResponse partition : topic.partitions()) {
-                TopicPartition topicPartition = new TopicPartition(topic.name(), partition.partitionIndex());
-                Errors error = Errors.forCode(partition.errorCode());
-                switch (error) {
-                    case NONE:
-                        if (!partition.oldStyleOffsets().isEmpty()) {
-                            // Handle v0 response with offsets
-                            long offset;
-                            if (partition.oldStyleOffsets().size() > 1) {
-                                future.raise(new IllegalStateException("Unexpected partitionData response of length " +
-                                        partition.oldStyleOffsets().size()));
-                                return;
-                            } else {
-                                offset = partition.oldStyleOffsets().get(0);
-                            }
-                            log.debug("Handling v0 ListOffsetResponse response for {}. Fetched offset {}",
-                                    topicPartition, offset);
-                            if (offset != ListOffsetsResponse.UNKNOWN_OFFSET) {
-                                ListOffsetData offsetData = new ListOffsetData(offset, null, Optional.empty());
-                                fetchedOffsets.put(topicPartition, offsetData);
-                            }
-                        } else {
-                            // Handle v1 and later response or v0 without offsets
-                            log.debug("Handling ListOffsetResponse response for {}. Fetched offset {}, timestamp {}",
-                                    topicPartition, partition.offset(), partition.timestamp());
-                            if (partition.offset() != ListOffsetsResponse.UNKNOWN_OFFSET) {
-                                Optional<Integer> leaderEpoch = (partition.leaderEpoch() == ListOffsetsResponse.UNKNOWN_EPOCH)
-                                        ? Optional.empty()
-                                        : Optional.of(partition.leaderEpoch());
-                                ListOffsetData offsetData = new ListOffsetData(partition.offset(), partition.timestamp(),
-                                        leaderEpoch);
-                                fetchedOffsets.put(topicPartition, offsetData);
-                            }
-                        }
-                        break;
-                    case UNSUPPORTED_FOR_MESSAGE_FORMAT:
-                        // The message format on the broker side is before 0.10.0, which means it does not
-                        // support timestamps. We treat this case the same as if we weren't able to find an
-                        // offset corresponding to the requested timestamp and leave it out of the result.
-                        log.debug("Cannot search by timestamp for partition {} because the message format version " +
-                                "is before 0.10.0", topicPartition);
-                        break;
-                    case NOT_LEADER_OR_FOLLOWER:
-                    case REPLICA_NOT_AVAILABLE:
-                    case KAFKA_STORAGE_ERROR:
-                    case OFFSET_NOT_AVAILABLE:
-                    case LEADER_NOT_AVAILABLE:
-                    case FENCED_LEADER_EPOCH:
-                    case UNKNOWN_LEADER_EPOCH:
-                        log.debug("Attempt to fetch offsets for partition {} failed due to {}, retrying.",
-                                topicPartition, error);
-                        partitionsToRetry.add(topicPartition);
-                        break;
-                    case UNKNOWN_TOPIC_OR_PARTITION:
-                        log.warn("Received unknown topic or partition error in ListOffset request for partition {}", topicPartition);
-                        partitionsToRetry.add(topicPartition);
-                        break;
-                    case TOPIC_AUTHORIZATION_FAILED:
-                        unauthorizedTopics.add(topicPartition.topic());
-                        break;
-                    default:
-                        log.warn("Attempt to fetch offsets for partition {} failed due to unexpected exception: {}, retrying.",
-                                topicPartition, error.message());
-                        partitionsToRetry.add(topicPartition);
-                }
-            }
-        }
-
-        if (!unauthorizedTopics.isEmpty())
-            future.raise(new TopicAuthorizationException(unauthorizedTopics));
-        else
-            future.complete(new ListOffsetResult(fetchedOffsets, partitionsToRetry));
-    }
-
-    static class ListOffsetResult {
-        private final Map<TopicPartition, ListOffsetData> fetchedOffsets;
-        private final Set<TopicPartition> partitionsToRetry;
-
-        ListOffsetResult(Map<TopicPartition, ListOffsetData> fetchedOffsets, Set<TopicPartition> partitionsNeedingRetry) {
-            this.fetchedOffsets = fetchedOffsets;
-            this.partitionsToRetry = partitionsNeedingRetry;
-        }
-
-        ListOffsetResult() {
-            this.fetchedOffsets = new HashMap<>();
-            this.partitionsToRetry = new HashSet<>();
+        try {
+            ListOffsetResult result = offsetFetcherUtils.handleListOffsetResponse(listOffsetsResponse);
+            future.complete(result);
+        } catch (RuntimeException e) {
+            future.raise(e);
         }
     }
+
+
 
     /**
      * If we have seen new metadata (as tracked by {@link org.apache.kafka.clients.Metadata#updateVersion()}), then
      * we should check that all the assignments have a valid position.
      */
     public void validatePositionsOnMetadataChange() {
-        int newMetadataUpdateVersion = metadata.updateVersion();
-        if (metadataUpdateVersion.getAndSet(newMetadataUpdateVersion) != newMetadataUpdateVersion) {
-            subscriptions.assignedPartitions().forEach(topicPartition -> {
-                ConsumerMetadata.LeaderAndEpoch leaderAndEpoch = metadata.currentLeader(topicPartition);
-                subscriptions.maybeValidatePositionForCurrentLeader(apiVersions, topicPartition, leaderAndEpoch);
-            });
-        }
+        offsetFetcherUtils.validatePositionsOnMetadataChange();
     }
 
     private Map<Node, Map<TopicPartition, FetchPosition>> regroupFetchPositionsByLeader(
@@ -701,17 +562,6 @@ public class OffsetFetcher {
                 .filter(entry -> entry.getValue().currentLeader.leader.isPresent())
                 .collect(Collectors.groupingBy(entry -> entry.getValue().currentLeader.leader.get(),
                         Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
-    }
-
-    private <T> Map<Node, Map<TopicPartition, T>> regroupPartitionMapByNode(Map<TopicPartition, T> partitionMap) {
-        return partitionMap.entrySet()
-                .stream()
-                .collect(Collectors.groupingBy(entry -> metadata.fetch().leaderFor(entry.getKey()),
-                        Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
-    }
-
-    private Set<String> topicsForPartitions(Collection<TopicPartition> partitions) {
-        return partitions.stream().map(TopicPartition::topic).collect(Collectors.toSet());
     }
 
 }
